Assignment 4- Code/requirements.txt
pinecone-client
langchain
langchain-openai
openai
fpdf2
fastapi
python-dotenv
google-cloud-storage
requests
serpapi
uvicorn
pytest==7.4.0
pytest-asyncio==0.21.1
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
pydantic[email]==2.4.2
google-search-results 
arxiv
langgraph
streamlit
